Detect interrupted game and restore
Moving to custom lists for this, and also add a menu at the start that can potentially skip the initialization phase. We also have to keep track of values like the current player (previously C), the top of the discard (previously L), and whether we are in the flip or play phase.
Unfortunately there are a few problems with this work so far:
- If the game is interrupted during first flip phase, only newly flipping players are considered for the "most points" comparison.
- If the game is interrupted when a player has flipped only one card, it won't be shown and they get to flip two more.
- If the game is interrupted after a player draws a card, that card is lost.
For the first problem, I think I will just unflip every card before the flip phase, and always start from player 1.
For the second problem, I don't really care. This does leave exploits technically.
And I also just realized I need to keep track of the draw position (previously D).